StockContainer class by merging contents into AnimalGroup class. Renamed stock_intf.cs to StockList.cs
